Function that chooses a puzzle
Added a function that selects one puzzle from the 10000 in the
"puzzles0_kaggle" file, for now it may only choose sudokus from that file,
so there is only one difficulty (easy).
Also it's only called when the app is initialized, so for a new puzzle you must restart the app.